Restructured Map file and modified the json writer
