+++ dsnake4-python/jsonTranslator.py
@@ -1,4 +1,5 @@
 from Map_Objects import *
+from Map import Map
 import json
 
 level_name = "Shop"
@@ -7,7 +8,7 @@ floor = 4
 level_map = Map()
 
 walls = []
-stairs = [{}]
+stairs = []
 
 for w in level_map.get_walls_at_floor(floor):
     direction = 'Down'
@@ -16,8 +17,8 @@ for w in level_map.get_walls_at_floor(floor):
 
     walls.append({
         "position": {
-            "x": w.start[0],
-            "y": w.start[1]
+            "x": w.start[0] + 1,
+            "y": w.start[1] + 1
         },
         "direction": direction,
         "removable": w.breakable,
@@ -25,5 +26,26 @@ for w in level_map.get_walls_at_floor(floor):
         "visible": True
     })
 
-with open(level_name + '.json', 'w') as outfile:
+for s in level_map.get_stairs_at_floor(floor):
+    if s.direction == (0, 1):
+        direction = 'Down'
+    elif s.direction == (0, -1):
+        direction = 'Up'
+    elif s.direction == (1, 0):
+        direction = 'Right'
+    elif s.direction == (-1, 0):
+        direction = 'Left'
+
+    stairs.append({
+        "position": {
+            "x": s.top_left[0] + 1,
+            "y": s.top_left[1] + 1
+        },
+        "exitDirection": direction,
+        "width": s.width,
+        "height": s.height,
+        "identifier": str(s.identifier)
+    })
+
+with open('dsnake4-python/json_data/' + level_name + '.json', 'w') as outfile:
     json.dump({"walls": walls, "stairs": stairs}, outfile)
