+++ dsnake4-python/json_data/Tropen.json
@@ -1,177 +1,1 @@
-{
-    "walls": [
-        {
-            "position": {
-                "x": 0,
-                "y": 0
-            },
-            "direction": "Right",
-            "removable": false,
-            "length": 104,
-            "visible": true
-        },
-        {
-            "position": {
-                "x": 104,
-                "y": 0
-            },
-            "direction": "Down",
-            "removable": false,
-            "length": 104,
-            "visible": true
-        },
-        {
-            "position": {
-                "x": 0,
-                "y": 59
-            },
-            "direction": "Right",
-            "removable": false,
-            "length": 104,
-            "visible": true
-        },
-        {
-            "position": {
-                "x": 0,
-                "y": 0
-            },
-            "direction": "Down",
-            "removable": false,
-            "length": 59,
-            "visible": true
-        },
-        {
-            "position": {
-                "x": 15,
-                "y": 0
-            },
-            "direction": "Down",
-            "removable": false,
-            "length": 59,
-            "visible": true
-        },
-        {
-            "position": {
-                "x": 90,
-                "y": 0
-            },
-            "direction": "Down",
-            "removable": false,
-            "length": 59,
-            "visible": true
-        },
-        {
-            "position": {
-                "x": 15,
-                "y": 15
-            },
-            "direction": "Right",
-            "removable": true,
-            "length": 25,
-            "visible": true
-        },
-        {
-            "position": {
-                "x": 45,
-                "y": 15
-            },
-            "direction": "Right",
-            "removable": true,
-            "length": 25,
-            "visible": true
-        },
-        {
-            "position": {
-                "x": 75,
-                "y": 15
-            },
-            "direction": "Right",
-            "removable": true,
-            "length": 15,
-            "visible": true
-        },
-        {
-            "position": {
-                "x": 40,
-                "y": 15
-            },
-            "direction": "Down",
-            "removable": true,
-            "length": 5,
-            "visible": true
-        },
-        {
-            "position": {
-                "x": 40,
-                "y": 24
-            },
-            "direction": "Down",
-            "removable": true,
-            "length": 18,
-            "visible": true
-        },
-        {
-            "position": {
-                "x": 40,
-                "y": 46
-            },
-            "direction": "Down",
-            "removable": true,
-            "length": 13,
-            "visible": true
-        },
-        {
-            "position": {
-                "x": 17,
-                "y": 37
-            },
-            "direction": "Right",
-            "removable": true,
-            "length": 23,
-            "visible": true
-        },
-        {
-            "position": {
-                "x": 53,
-                "y": 3
-            },
-            "direction": "Down",
-            "removable": true,
-            "length": 12,
-            "visible": true
-        },
-        {
-            "position": {
-                "x": 65,
-                "y": 35
-            },
-            "direction": "Right",
-            "removable": true,
-            "length": 24,
-            "visible": true
-        },
-        {
-            "position": {
-                "x": 65,
-                "y": 35
-            },
-            "direction": "Down",
-            "removable": true,
-            "length": 9,
-            "visible": true
-        },
-        {
-            "position": {
-                "x": 65,
-                "y": 48
-            },
-            "direction": "Down",
-            "removable": true,
-            "length": 11,
-            "visible": true
-        }
-    ],
-    "stairs": [
-        {}
-    ]
-}
+{"walls": [{"position": {"x": 1, "y": 1}, "direction": "Right", "removable": false, "length": 104, "visible": true}, {"position": {"x": 105, "y": 1}, "direction": "Down", "removable": false, "length": 59, "visible": true}, {"position": {"x": 1, "y": 60}, "direction": "Right", "removable": false, "length": 50, "visible": true}, {"position": {"x": 56, "y": 60}, "direction": "Right", "removable": false, "length": 50, "visible": true}, {"position": {"x": 1, "y": 1}, "direction": "Down", "removable": false, "length": 59, "visible": true}, {"position": {"x": 16, "y": 1}, "direction": "Down", "removable": false, "length": 59, "visible": true}, {"position": {"x": 91, "y": 1}, "direction": "Down", "removable": false, "length": 59, "visible": true}, {"position": {"x": 16, "y": 16}, "direction": "Right", "removable": true, "length": 25, "visible": true}, {"position": {"x": 46, "y": 16}, "direction": "Right", "removable": true, "length": 25, "visible": true}, {"position": {"x": 76, "y": 16}, "direction": "Right", "removable": true, "length": 15, "visible": true}, {"position": {"x": 41, "y": 16}, "direction": "Down", "removable": true, "length": 5, "visible": true}, {"position": {"x": 41, "y": 25}, "direction": "Down", "removable": true, "length": 18, "visible": true}, {"position": {"x": 41, "y": 47}, "direction": "Down", "removable": true, "length": 13, "visible": true}, {"position": {"x": 18, "y": 38}, "direction": "Right", "removable": true, "length": 23, "visible": true}, {"position": {"x": 54, "y": 4}, "direction": "Down", "removable": true, "length": 12, "visible": true}, {"position": {"x": 66, "y": 36}, "direction": "Right", "removable": true, "length": 24, "visible": true}, {"position": {"x": 66, "y": 36}, "direction": "Down", "removable": true, "length": 9, "visible": true}, {"position": {"x": 66, "y": 49}, "direction": "Down", "removable": true, "length": 11, "visible": true}], "stairs": [{"position": {"x": 51, "y": 60}, "exitDirection": "Up", "width": 5, "height": 1, "identifier": "5"}]}
